+++ src/scripts/components/tree/tree-node.vue
@@ -71,7 +71,7 @@
 
         <label
           class="mdc-tree-node__label"
-          @click.self.prevent="
+          @click.prevent="
             treeData.multiple ? handleCheck(nodeData) : handleSelect(nodeData)
           "
         >
@@ -133,7 +133,9 @@ export default {
       MdcTree.onExpand(this.treeData, item);
     },
     handleSelect(item) {
-      MdcTree.onSelect(this.treeData, item);
+      if (!item.disabled) {
+        MdcTree.onSelect(this.treeData, item);
+      }
     },
     handleCheck(item) {
       if (!item.disabled) {