+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 > :tada: [`balm-ui`](https://github.com/balmjs/balm-ui)(v9+) supports for Vue 3
 
+## v8.63.1 / 2023-11-07
+
+### Chore
+
+- `<ui-form>`, `<ui-form-field>`: optimize initialization
+
+## v8.63.0 / 2023-10-31
+
+### Bug Fixes
+
+- `<ui-side-sheet>`: fix individual bug
+
+## v8.62.1 / 2023-09-21
+
+### Bug Fixes
+
+- `<ui-tree>`: fix prevent click for custom label
+
 ## v8.62.0 / 2023-09-15
 
 ### Features